% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d}
\alias{d}
\title{Example Network Data 1}
\format{
## `d`
A data frame with 53 rows and 9 columns:
\describe{
  \item{Cluster.ID}{lipid class ID}
  \item{label.saturation}{lipid class stratified by saturation}
  \item{p.value}{raw p-value}
  \item{FDR}{FDR-adjusted p-value}
  \item{n}{cluster size}
  \item{Increased}{number of increased lipids within cluster}
  \item{Decreased}{number of decreased lipids within cluster}
  \item{Ratio}{proportion of increased/decreased lipids within cluster}
  \item{label}{lipid class label}
  ...
}
}
\source{
HiVE v1.10
}
\usage{
d
}
\description{
Example data for constructing HiVE network plots
based on lipidomics data.
}
\keyword{datasets}
