% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d2}
\alias{d2}
\title{Example Network Data 2}
\format{
## `d2`
A data frame with 53 rows and 9 columns:
\describe{
  \item{Type}{sample type}
  \item{Name}{lipid name}
  \item{Comparison}{group comparison}
  \item{p.adj}{adjusted p-value}
  \item{FDR}{FDR adjusted p-value}
  \item{Comparison.fc}{fold change comparison direction}
  \item{Log2FC}{fold change}
  \item{label}{lipid class label}
  \item{n}{number of lipids in class}
  ...
}
}
\source{
HiVE v1.10
}
\usage{
d2
}
\description{
Example data for constructing individual lipid-level network plots
based on lipidomics data.
}
\keyword{datasets}
