\name{DFS}
\alias{DFS_cat}
\alias{DFS_dir_create}
\alias{DFS_dir_exists}
\alias{DFS_dir_remove}
\alias{DFS_file_exists}
\alias{DFS_format}
\alias{DFS_list}
\alias{DFS_put}
\alias{DFS_put_object}
\alias{DFS_tail}

\title{Hadoop Distributed File System}
\description{
  
  Functions providing high-level access to the Hadoop Distributed File System (HDFS).
  
}
\usage{

}
\arguments{
  \item{henv}{Hadoop local environment.}
}
\details{

}
\value{

}
\references{
  Apache Hadoop core
  (\url{http://hadoop.apache.org/core/}).  
}

\author{Stefan Theussl}
%\seealso{
%  \pkg{RHIPE}
%  \pkg{HadoopStreaming}
%}
\examples{

}
%%\keyword{hpc}
