% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{edgeoxy}
\alias{edgeoxy}
\title{Oxylipin Subnetwork Edges}
\format{
## `edgeoxy`
A data frame with 56 rows and 8 columns:
\describe{
  \item{from}{lipid precursor}
  \item{to}{lipid product}
  \item{synthesis.pathway}{lipid synthesis pathway information}
  \item{id.enzyme}{representative enzyme name}
  \item{id.gene}{representative gene name}
  \item{hgnc.symbols}{associated genes}
  \item{SetID}{lipid set identity}
  \item{Level}{network level}
  \item{major.enzyme}{primary oxylipin enzyme pathway}
  \item{start.fa}{Starting fatty acid name}
  ...
}
}
\source{
HiVE v1.10
}
\usage{
edgeoxy
}
\description{
Edge metadata for constructing a HiVE oxylipin subnetwork.
}
\keyword{datasets}
