% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01lib.R
\name{gen_theme}
\alias{gen_theme}
\title{Generic Plot Theme}
\usage{
gen_theme()
}
\value{
ggplot2 theme parameters to replace default plot theme.
}
\description{
General plotting theme.
}
\examples{

# sc_theme1()

}
