% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiVE-package.R
\docType{package}
\name{HiVE-package}
\alias{HiVE}
\alias{HiVE-package}
\title{HiVE: High-dimensional lipid Visualization and Enrichment analysis (HiVE)}
\description{
This package provides tools for comprehensive visualization and analysis of multiomics data characterizing lipids. HiVE integrates underlying knowledge of curated lipid metabolic pathways for quick mapping and analysis of lipidomics data onto a predefined network structure. Flexible data formats supported by HiVE also enable simultaneous or individual mapping of complementary proteomics or transcriptomics datasets containing lipid metabolizing enzymes or lipid-related genes.
}
\keyword{internal}
