% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04netanno.r
\name{hive_anno}
\alias{hive_anno}
\title{Lipid Network Annotation Plot}
\usage{
hive_anno(level = 0, enr_obj)
}
\arguments{
\item{level}{Base level to plot (either 0 or 1). Level 1 contains
finer class-based separation while 0 (default) plots higher level
lipid classes.}

\item{enr_obj}{An enrichment object including lipid class labels
and the number of lipids within each class. The labels must match
nodes from the HiVE base network nodes to properly function. Example
data are included for formatting help.}
}
\value{
A ggplot2 object plotting the HiVE base network.
}
\description{
Extension of hive_base() for adding annotation count information
to a base HiVE network. Node sizes are proportional to the number
of individual lipids present in the dataset for each class. See
HiVE example data for more details.
}
\examples{

# ## Annotation count network
# hive_anno(enr_obj = d)

}
