% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03netbase.r
\name{hive_base}
\alias{hive_base}
\title{Lipid Network Plot}
\usage{
hive_base()
}
\value{
A ggplot2 object plotting the HiVE base network.
}
\description{
Generates a network plot using known lipid metabolism pathways
to visualize class-based enrichment statistics. Uses a manually
curated collection of pathways as a reference. Lipid class
annotations assigned by this reference enable mapping of any
untargeted or targeted lipidomics dataset onto the network.
}
\examples{

# ## Base network
# hive_base()

}
