% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05netenrch.r
\name{hive_enrch}
\alias{hive_enrch}
\title{Lipid Network Enrichment Plot}
\usage{
hive_enrch(
  type = "base",
  lab_a = 0.5,
  enr_obj,
  col_lab = "Class",
  col_sat = "label.saturation"
)
}
\arguments{
\item{type}{Network type to plot; current lipid networks available are
"base" (Default) and "oxylipins", which is a more detailed subnetwork
of HiVE base.}

\item{lab_a}{Edge label alpha value.}

\item{enr_obj}{An enrichment object including lipid class labels
and the number of lipids within each class. The labels must match
nodes from the HiVE base network nodes to properly function. Example
data are included for formatting help.}

\item{col_lab}{Compound class column name from the input enrichment
object.}

\item{col_sat}{Column name that indicates the saturation degree
of each lipid class included in the input enrichment object. Use the
same column for both col_lab and col_sat if compound classes split
by saturation degree are not available.}
}
\value{
A ggplot2 object plotting a HiVE enrichment network.
}
\description{
Extension of hive_base() for plotting enrichment results onto
a base HiVE network. Node sizes are proportional to the number
of individual lipids present in the dataset for each class, node
colors represent the proportion of increased/decreased lipids
within a specific class, and yellow boxes indicate classes that
are significantly enriched in the selected comparison. See
HiVE example data for more details. Please note that if plotting
data onto a subnetwork, additional data formatting is necessary as
these data are usually not provided as enrichment objects.
}
\examples{

# ## Annotation count network
# hive_enrch(enr_obj = d)

}
