% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07netflux.r
\name{hive_flux}
\alias{hive_flux}
\title{Lipid Network Flux Plot}
\usage{
hive_flux(
  type = "base",
  lab_a = 0.5,
  stat_obj,
  col_fc = "Log2FC",
  col_p = "p adj",
  col_rat = "Name"
)
}
\arguments{
\item{type}{Network type to plot; current lipid networks available are
"base" (Default) and "oxylipins", which is a more detailed subnetwork
of HiVE base.}

\item{lab_a}{Edge label alpha value.}

\item{stat_obj}{A data frame of statistical results containing
fold changes and p-values run by ms_stat_uni using the "HiVE" method.}

\item{col_fc}{Fold change column name.}

\item{col_p}{p value column name.}

\item{col_rat}{Column containing associated lipid class ratios.}
}
\value{
A ggplot2 object plotting a HiVE flux network.
}
\description{
Extension of hive_base() for plotting ratio analysis results onto
a base HiVE network. Solid edges indicate statistically significant
ratios, edge colors represent the log2 fold change of the specified
comparison, and grey edges indicate that the ratio between two lipid
classes was not calculated. See HiVE example data for more details.
}
\examples{

# ## Annotation count network
# hive_flux(enr_obj = d)

}
