% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06netindv.r
\name{hive_indv}
\alias{hive_indv}
\title{Lipid Network Individual Data Plot}
\usage{
hive_indv(
  type = "oxylipins",
  stat_obj,
  col_fc = "Log2FC",
  col_p = "p.adj",
  col_lab = "label",
  col_ct = "CellType",
  plot_mode = "lipidomics",
  gex_type = "GEX",
  gex_filt = TRUE,
  tx_size_major = 4,
  tx_size_minor = 3,
  show_minor_lab = TRUE,
  set_incr = NULL
)
}
\arguments{
\item{type}{Network type to plot; current lipid networks available are
"base" and "oxylipins" (Default), which is a more detailed subnetwork
of HiVE base.}

\item{stat_obj}{A data frame of statistical results containing
fold changes and p-values or a Seurat object. The labels must match
nodes from the HiVE base network nodes to properly function. Example
data are included for formatting help. Can also plot differential
gene expression or general expression data if desired.}

\item{col_fc}{Fold change column name.}

\item{col_p}{p value column name.}

\item{col_lab}{Label column for mapping input data onto HiVE nodes.}

\item{col_ct}{Cell type column name (only applicable if plot_mode is
"transcriptomics").}

\item{plot_mode}{Plot mode; select "lipidomics" for plotting statistical
results for lipidomics data or "transcriptomics" for plotting gene expression
or DGEA results.}

\item{gex_type}{Gene expression data type; either "GEX" or "DGEA". Can toggle
gex_filt to TRUE to plot only the primary gene between two nodes or FALSE to
plot expression/fold changes for all relevant genes between two nodes.}

\item{gex_filt}{Plot only the primary gene between two nodes?}

\item{tx_size_major}{Major node label text size.}

\item{tx_size_minor}{Minor node label text size.}

\item{show_minor_lab}{Show minor node labels?}

\item{set_incr}{Set the step distance between expression plot points. The
distance is automatically scaled to the number of points if NULL.}
}
\value{
A ggplot2 object plotting individual annotation-level
data onto the HiVE base network.
}
\description{
Extension of hive_base() for plotting individual abundance/expression
results onto a base HiVE network. Node/edge colors represent the
fold change for the specified comparison for individual lipid
species/genes. See HiVE example data for more details.
Note that these comparisons are most effective when plotted on
a subnetwork.
}
\examples{

# ## Annotation count network
# hive_indv(enr_obj = d)

}
