% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09netsplit.r
\name{hive_panel_edge}
\alias{hive_panel_edge}
\title{Lipid Network Edge Panel}
\usage{
hive_panel_edge(
  stat_obj,
  name_edge,
  col_ct = "CellType",
  edge_type = "GEX",
  tx_size_col = 4,
  tx_size_row = 3,
  nudge_col = 0.025,
  nudge_row = -0.01,
  p_size = 8,
  set_incr = 0.03,
  lab_col_bound = 0.055,
  lab_row_bound = 0.045
)
}
\arguments{
\item{stat_obj}{A data frame of statistical results containing
fold changes and p-values or a Seurat object. The labels must match
nodes from the HiVE base network nodes to properly function. Example
data are included for formatting help. Can also plot differential
gene expression or general expression data if desired.}

\item{name_edge}{Vector of genes to plot.}

\item{col_ct}{Cell type column name.}

\item{edge_type}{Either "GEX" or "DGEA".}

\item{tx_size_col}{Column label text size.}

\item{tx_size_row}{Row label text size.}

\item{nudge_col}{Nudge column labels along y-axis.}

\item{nudge_row}{Nudge row labels along x-axis.}

\item{p_size}{Plot point size.}

\item{set_incr}{Step size between plot points.}

\item{lab_col_bound}{Boundary box y limit.}

\item{lab_row_bound}{Boundary box x limit.}
}
\value{
A ggplot2 object plotting individual annotation-level
data onto the HiVE base network.
}
\description{
Extension of hive_mult() for isolating edge panels for custom
figure design. Edge box colors represent the
fold change for the specified comparison for individual lipid
genes. See HiVE example data for more details.
Note that this function is intended for manually plotting data
onto a network generated by hive_base().
}
\examples{

# ## HiVE edge panel
# hive_panel_edge(
#   stat_obj = d,
#   name_edge = c("ALOX5", "ALOX12", "ALOX15")
# )

}
