% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09netsplit.r
\name{hive_panel_node}
\alias{hive_panel_node}
\title{Lipid Network Circular Node Panel}
\usage{
hive_panel_node(
  stat_obj,
  name_node,
  col_fc = "Log2FC",
  col_p = "p.adj",
  col_lab = "label",
  tx_size_major = 4,
  tx_size_minor = 3,
  p_size = 5
)
}
\arguments{
\item{stat_obj}{A data frame of statistical results containing
fold changes and p-values or a Seurat object. The labels must match
nodes from the HiVE base network nodes to properly function. Example
data are included for formatting help. Can also plot differential
gene expression or general expression data if desired.}

\item{name_node}{Name of the node to plot.}

\item{col_fc}{Fold change column name.}

\item{col_p}{p value column name.}

\item{col_lab}{Label column for mapping input data onto HiVE nodes.}

\item{tx_size_major}{Major node label text size.}

\item{tx_size_minor}{Minor node label text size.}

\item{p_size}{Minor node point size.}
}
\value{
A ggplot2 object plotting circular node points for a specific
HiVE node.
}
\description{
Extension of hive_mult() for isolating node panels for custom
figure design. Node colors represent the
fold change for the specified comparison for individual lipid
species. See HiVE example data for more details.
Note that this function is intended for manually plotting data
onto a network generated by hive_base().
}
\examples{

# ## Annotation count network
# hive_panel_node(stat_obj = d1, name_node = "HETE")

}
