% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nedge}
\alias{nedge}
\title{Lipid Network Edges}
\format{
## `nedge`
A data frame with 163 rows and 8 columns:
\describe{
  \item{from}{lipid precursor}
  \item{to}{lipid product}
  \item{synthesis.pathway}{lipid synthesis pathway information}
  \item{leading.name}{representative gene/enzyme name}
  \item{enzyme.id}{unique gene/enzyme identity code}
  \item{hgnc.symbols}{associated genes}
  \item{SetID}{lipid set identity}
  \item{Level}{network level}
  ...
}
}
\source{
HiVE v1.02
}
\usage{
nedge
}
\description{
Edge metadata for constructing a HiVE lipid network.
}
\keyword{datasets}
