% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01lib.R
\name{net_theme}
\alias{net_theme}
\title{Network Plot Theme}
\usage{
net_theme(leg = c(0.9, 0.92))
}
\arguments{
\item{leg}{X and Y coordinates of the desired legend position,
provided as a vector.}
}
\value{
ggplot2 theme parameters to replace default plot theme.
}
\description{
General network plot theme.
}
\examples{

# net_theme(leg = c(0.9, 0.1))

}
