% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{netgene}
\alias{netgene}
\title{Lipid Gene List}
\format{
## `netgene`
A data frame with 299 rows and 5 columns:
\describe{
  \item{Gene}{gene name}
  \item{SetID}{lipid pathway name}
  \item{abbv}{unique pathway code}
  \item{RefType}{Reference type}
  \item{Source}{DOI/PMCID if applicable}
  ...
}
}
\source{
HiVE v1.10
}
\usage{
netgene
}
\description{
Reference lipid gene list.
}
\keyword{datasets}
