% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nnode}
\alias{nnode}
\title{Lipid Network Nodes}
\format{
## `nnode`
A data frame with 151 rows and 5 columns:
\describe{
  \item{Node}{shorthand node name}
  \item{Name}{full node name}
  \item{synthesis.pathway}{lipid synthesis pathway information}
  \item{SetID}{lipid set identity}
  \item{Level}{network level}
  ...
}
}
\source{
HiVE v1.02
}
\usage{
nnode
}
\description{
Node metadata for constructing a HiVE lipid network.
}
\keyword{datasets}
