% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nodeoxy}
\alias{nodeoxy}
\title{Oxylipin Subnetwork Nodes}
\format{
## `nodeoxy`
A data frame with 52 rows and 10 columns:
\describe{
  \item{Node}{shorthand node name}
  \item{Root}{root node}
  \item{Name}{full node name}
  \item{Synonyms}{node synonyms}
  \item{Form}{fatty acid chain length and saturation}
  \item{synthesis.pathway}{lipid synthesis pathway information}
  \item{enzyme.class}{major enzyme class}
  \item{Level}{network level}
  \item{hive_x}{node x coordinate}
  \item{hive_y}{node y coordinate}
  ...
}
}
\source{
HiVE v1.10
}
\usage{
nodeoxy
}
\description{
Node metadata for constructing a HiVE oxylipin subnetwork.
}
\keyword{datasets}
