% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02CDK_extension.R
\name{view_structure}
\alias{view_structure}
\title{Generate Structure from SMILES}
\usage{
view_structure(smi, sav = FALSE, title1 = NULL)
}
\arguments{
\item{smi}{Input SMILES code.}

\item{sav}{Logical indicating whether the image should be saved
to the disk.}

\item{title1}{If sav is TRUE, provide a file name for saving
the image.}
}
\value{
Raster image generated from SMILEs code.
}
\description{
Creates and/or saves a raster image from a SMILES
code using the R CDK framework.
}
\examples{

 # view_structure(smi = "C(C(=O)[O-])C(CC(=O)[O-])(C(=O)[O-])O")

}
