% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04netanno.r
\name{hive_anno}
\alias{hive_anno}
\title{Lipid Network Annotation Plot}
\usage{
hive_anno(type = "base", lab_a = 0.5, enr_obj)
}
\arguments{
\item{type}{Network type to plot; current lipid networks available are
"base" (Default) and "oxylipins", which is a more detailed subnetwork
of HiVE base.}

\item{lab_a}{Alpha value for enzyme names (default is transparent).}

\item{enr_obj}{An enrichment object including lipid class labels
and the number of lipids within each class. The labels must match
nodes from the HiVE base network nodes to properly function. Example
data are included for formatting help. Univariate statistical results
can also be provided as input for annotating subnetwork plots.}
}
\value{
A ggplot2 object plotting the HiVE annotation network.
}
\description{
Extension of hive_base() for adding annotation count information
to a base HiVE network. Node sizes are proportional to the number
of individual lipids present in the dataset for each class. See
HiVE example data for more details.
}
\examples{

# ## Annotation count network
# hive_anno(enr_obj = d)

}
