% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03netbase.r
\name{hive_base}
\alias{hive_base}
\title{Lipid Network Plot}
\usage{
hive_base(type = "base", lab_a = 0, filt_pw = NULL)
}
\arguments{
\item{type}{Network type to plot; current lipid networks available are
"base" (Default) and "oxylipins", which is a more detailed subnetwork
of HiVE base.}

\item{lab_a}{Alpha value for enzyme names (default is transparent).}

\item{filt_pw}{Filter by a specific pathway present in either the "base"
or "oxylipins" networks.}
}
\value{
A ggplot2 object plotting the HiVE base network.
}
\description{
Generates a network plot using known lipid metabolism pathways
to visualize class-based enrichment statistics. Uses a manually
curated collection of pathways as a reference. Lipid class
annotations assigned by this reference enable mapping of any
untargeted or targeted lipidomics dataset onto the network.
}
\examples{

# ## Base network
# hive_base()

}
