% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05netenrch.r
\name{hive_enrch}
\alias{hive_enrch}
\title{Lipid Network Enrichment Plot}
\usage{
hive_enrch(level = 0, enr_obj, comp_sat = TRUE)
}
\arguments{
\item{level}{Base level to plot (either 0 or 1). Level 1 contains
finer class-based separation while 0 (default) plots higher level
lipid classes.}

\item{enr_obj}{An enrichment object including lipid class labels
and the number of lipids within each class. The labels must match
nodes from the HiVE base network nodes to properly function. Example
data are included for formatting help.}

\item{comp_sat}{Compare enrichment results split by lipid saturation.
This is currently the only option but more will be supported in
furture package versions.}
}
\value{
A ggplot2 object plotting a HiVE enrichment network.
}
\description{
Extension of hive_base() for plotting enrichment results onto
a base HiVE network. Node sizes are proportional to the number
of individual lipids present in the dataset for each class, node
colors represent the proportion of increased/decreased lipids
within a specific class, and yellow boxes indicate classes that
are significantly enriched in the selected comparison. See
HiVE example data for more details.
}
\examples{

# ## Annotation count network
# hive_enrch(enr_obj = d)

}
